split off embed and tokenize functions
